- Added a dormant service for checking on resolver
The rationale for not making this service enabled by default is that I
can envision a situation when somebody deploys an instance into a
network with no DNS resolution available.  The unit introduced with
this commit is unconditionally waits for the DNS resolution to be
working.  If you need to ensure that the resolver is up and is ready
to use before network-online.target is reached, then enable the
service with "systemctl enable systemd-resolved-wait-online".